moved fetch_paper_metadata.py and doi_resolve.py to sources, exposed commands from fetch_paper to init
